+++ test/libsolidity/syntaxTests/iceRegressionTests/declarationUnaryTuple/declaration_bitnot_tuple.sol
@@ -1,0 +1,9 @@
+contract C
+{
+    function f() public
+    {
+        int x = ~(0,);
+    }
+}
+// ----
+// TypeError: (60-64): Tuple component cannot be empty.