+++ Changelog.md
@@ -14,6 +14,7 @@ Bugfixes:
  * Type Checker: Disallow ``virtual`` and ``override`` for constructors.
  * Type Checker: Fix several internal errors by performing size and recursiveness checks of types before the full type checking.
  * Type Checker: Fix internal error when assigning to empty tuples.
+ * Type Checker: Fix internal error when applying unary operators to tuples with empty components.
  * Type Checker: Perform recursiveness check on structs declared at the file level.
  * Standard Json Input: Fix error when using prefix ``file://`` in the field ``urls``.